fix bug

+++ Source/Engine/Source/ING/ECS/Component/Id/Id.h
@@ -13,7 +13,7 @@ namespace ING {
 
 	namespace ECS {
 
-		using ComponentId = unsigned int;
+		using ComponentId = unsigned long;
 
 	}
